# Speed up packet dedup and fix benches