#### Problems
- Packet deduping is slow for larger packets but we always add the full 1232 data buffer to the bloom filter
- It's slow to iterate over the bloom filter keys twice (once for `contains` and once for `add`)

#### Summary of Changes
- Only add the used data buffer portion of a packet to the bloom filter
- Modify `AtomicBloom:add` to return whether the item was newly added or not

Before
```
test bench_dedup_diff_big_packets   ... bench:  20,010,316 ns/iter (+/- 1,623,356)
test bench_dedup_diff_small_packets ... bench:  19,772,433 ns/iter (+/- 1,470,336)
test bench_dedup_same_big_packets   ... bench:  17,869,979 ns/iter (+/- 1,158,488)
test bench_dedup_same_small_packets ... bench:  18,181,612 ns/iter (+/- 2,449,875)
```

After
```
test bench_dedup_diff_big_packets   ... bench:  15,266,433 ns/iter (+/- 1,096,914)
test bench_dedup_diff_small_packets ... bench:   1,943,839 ns/iter (+/- 239,305)
test bench_dedup_same_big_packets   ... bench:  16,299,237 ns/iter (+/- 570,619)
test bench_dedup_same_small_packets ... bench:   3,363,462 ns/iter (+/- 2,007,400)
```

Fixes #